lightning: move OpenError to channels
