+++ packages/lightning/lib/channels/ChannelManager.ts
@@ -7,8 +7,8 @@ import { IChannelLogic } from "./IChannelLogic";
 import { IPeer } from "../Peer";
 import { Result } from "../Result";
 import { OpenChannelRequest } from "./OpenChannelRequest";
-import { OpeningError } from "./states/opening/OpeningError";
-import { OpeningErrorType } from "./states/opening/OpeningErrorType";
+import { OpeningError } from "./OpeningError";
+import { OpeningErrorType } from "./OpeningErrorType";
 import { PeerState } from "../PeerState";
 import { AcceptChannelMessage } from "../messages/AcceptChannelMessage";
 import { IStateMachine } from "./IStateMachine";
