+++ packages/lightning/lib/channels/Helpers.ts
@@ -17,8 +17,8 @@ import { Result } from "../Result";
 import { Channel } from "./Channel";
 import { IChannelWallet } from "./IChannelWallet";
 import { OpenChannelRequest } from "./OpenChannelRequest";
-import { OpeningError } from "./states/opening/OpeningError";
-import { OpeningErrorType } from "./states/opening/OpeningErrorType";
+import { OpeningError } from "./OpeningError";
+import { OpeningErrorType } from "./OpeningErrorType";
 import { OpenChannelMessage } from "../messages/OpenChannelMessage";
 import { IChannelLogic } from "./IChannelLogic";
 import { AcceptChannelMessage } from "../messages/AcceptChannelMessage";
