+++ packages/lightning/lib/channels/OpeningErrorType.ts
